# hugoproxy-monitoring/proxy/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для поиска адресов и геокодирования",
        "title": "Геосервис API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/address/geocode": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Получение адреса по координатам",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "address"
                ],
                "summary": "Геокодирование адреса",
                "parameters": [
                    {
                        "description": "Координаты для геокодирования",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.GeocodeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ с найденными адресами",
                        "schema": {
                            "$ref": "#/definitions/service.GeocodeResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос",
                        "schema": {
                            "$ref": "#/definitions/responder.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/responder.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/responder.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/address/search": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Поиск адреса по строке запроса",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "address"
                ],
                "summary": "Поиск адреса",
                "parameters": [
                    {
                        "description": "Поисковый запрос",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.SearchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ с найденными адресами",
                        "schema": {
                            "$ref": "#/definitions/service.SearchResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос",
                        "schema": {
                            "$ref": "#/definitions/responder.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/responder.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/responder.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "responder.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "service.Address": {
            "description": "Информация об адресе",
            "type": "object",
            "properties": {
                "city": {
                    "description": "Название города",
                    "type": "string",
                    "example": "Москва"
                },
                "house": {
                    "description": "Номер дома",
                    "type": "string",
                    "example": "11"
                },
                "lat": {
                    "description": "Широта",
                    "type": "string",
                    "example": "55.7558"
                },
                "lon": {
                    "description": "Долгота",
                    "type": "string",
                    "example": "37.6173"
                },
                "street": {
                    "description": "Название улицы",
                    "type": "string",
                    "example": "Ленина"
                }
            }
        },
        "service.GeocodeRequest": {
            "description": "Запрос для геокодирования координат",
            "type": "object",
            "properties": {
                "lat": {
                    "description": "Широта",
                    "type": "string",
                    "example": "55.7558"
                },
                "lng": {
                    "description": "Долгота",
                    "type": "string",
                    "example": "37.6173"
                }
            }
        },
        "service.GeocodeResponse": {
            "description": "Ответ с геокодированными адресами",
            "type": "object",
            "properties": {
                "addresses": {
                    "description": "Список найденных адресов",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/service.Address"
                    }
                }
            }
        },
        "service.SearchRequest": {
            "description": "Запрос для поиска адреса",
            "type": "object",
            "properties": {
                "query": {
                    "description": "Поисковый запрос (город, улица, дом)",
                    "type": "string",
                    "example": "Москва Ленина 11"
                }
            }
        },
        "service.SearchResponse": {
            "description": "Ответ с найденными адресами",
            "type": "object",
            "properties": {
                "addresses": {
                    "description": "Список найденных адресов",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/service.Address"
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}